fix: Don't be paralyzed by overweight
`overweight` is a condition when Junkie allocates more memory than a given
limit. It then refrain from allocating some kind of data structures,
especially new parsers. But then, it did not run any more in the code
that timeout unused parsers.
This make things simpler : whenever we are in overweight the timeouter
thread is allowed to timeout any parsers.

Additionally, outputs an error message when entering overweight, and
allocate nicely for more object types.