new: TCP now synchronize one direction with regard to the other
Before, TCP streams were reordered in both directions individually, which
confused some parsers (for instance, HTTP parser may see an answer before
the query. Now, TCP uses the acknowledgment number to put on hold segments
that ack something we have not received yet.

Of course if your junkie is in bridge this problem never hit you, but when
the traffic is mirrored to your junkie that's another story.